fix: moving manifest list code to different modules
